ELF-oracle (gcc 2.7.2.3) fixes, all PE-oracle (MSVC6) regression-checked:
  intermediate `v2` cache variable for source->scriptstack, letting gcc's
  own CSE reproduce real's addressing; reordered floatvalue assignment
  last in evalint. Both byte-identical MATCH.
- PS_ExpectTokenType (l_script.c): removed invented `v6` cache of
  token->type — after the initial equality check, real reuses the
  register already holding `type` at every later comparison rather than
  re-reading token->type from memory, an optimization the cached variable
- Characteristic_Float (be_ai_char.c): applied sibling
  Characteristic_Integer's cached-pointer style. Improved 13->9
- Intersection (be_ai_move.c): fixed multiply operand order (det, c1,
- BotSwimInDirection (be_ai_move.c): replaced an overly-cautious
  int[3]+bitcast workaround with Q3A's plain vec3_t+VectorCopy (a
- AAS_NearbySolidOrGap (be_aas_reach.c): reused one named `areanum`
  variable across all 3 AAS_PointAreaNum call sites instead of leaving
  2 anonymous. Improved 12->6 insn_diffs (residual is a branch-layout
  floor).
  declared `int time` instead of `float time`, forcing an invented
  SLODWORD() bit-reinterpretation cast at the AAS_StartFrame call site
  and GP-register codegen instead of the FPU store real uses. Fixed both
  signatures (every declaration) to `float time` matching Q3A.
  improved 19->15. AAS_ContinueInitReachability unchanged (param unused
  inside; fix was only needed for the caller).

Investigated, left unchanged:
  source-level lever found after 3 distinct experiments (one, explicit
  zero-init, made it worse).
- PC_OperatorPriority (l_precomp.c): a case-label reorder achieves ELF
  byte-identical MATCH but regresses MSVC6's jump-table block layout for
  the same function (6 byte diffs) -- this is the third time this exact,
  already-documented conflict (see asm_matching_project.md, "Some
  switch-case block orderings are irreconcilable between gcc272 and
  MSVC6") has been independently re-derived and had to be reverted.
  Bisected this time: neither half of the two-part reorder is
  individually PE-free either. Reverted to the original PE-safe case
  order; ELF stays at its known floor of 10 insn_diffs.

sweep (not just per-function spot checks), with a second clean-baseline
control rebuild to positively distinguish genuine regressions from
BotChooseInitialChatMessage -- both differ even between two identical
clean rebuilds, confirmed pre-existing and unrelated). Final check: fresh
control build vs this batch's build, zero differences anywhere in the
~798-row sweep.

Native `make clean && make` verified clean on both botlib and game.
changes to record).

Source-commit: 923e85581c91f4389430a2ad919ec252c7e1f732

+++ botlib/be_aas_main.h
@@ -18,7 +18,7 @@ extern aas_link_t **aasentity_arealinks;
 extern bsp_link_t **aasentity_bsplinks;
 extern aas_world_t aasworld;
 
-int __cdecl AAS_ContinueInit(int time);
+int __cdecl AAS_ContinueInit(float time);
 int AAS_Error(char *Format, ...);
 char *__cdecl AAS_ImageFromIndex(int index);
 int __cdecl AAS_IndexFromImage(char *String2);
